Even more silencing of compiler warnings

+++ source/exhumed/src/status.cpp
@@ -827,7 +827,7 @@ void DrawStatus()
         sprintf(coordBuf, "X %d", (int)sprite[nSprite].x);
         printext(x, nViewTop + 1, coordBuf, kTile159, 255);
 
-        sprintf(coordBuf, "Y %d", sprite[nSprite].y);
+        sprintf(coordBuf, "Y %d", (int)sprite[nSprite].y);
         printext(x, nViewTop + 10, coordBuf, kTile159, 255);
     }
